+++ mod/usb/usb.h
@@ -43,18 +43,18 @@ typedef enum USB_STATUS
 /* transmits bytes over USB */
 USB_STATUS usb_transmit 
 	(
-    void*   tx_data_ptr , /* Data to be sent       */	
-	size_t  tx_data_size, /* Size of transmit data */ 
-	uint8_t timeout       /* UART timeout          */
+    void*    tx_data_ptr , /* Data to be sent       */	
+	size_t   tx_data_size, /* Size of transmit data */ 
+	uint32_t timeout       /* UART timeout          */
 	);
 
 
 /* Receives bytes from the USB port */
 USB_STATUS usb_receive 
 	(
-	void*   rx_data_ptr , /* Buffer to export data to        */
-	size_t  rx_data_size, /* Size of the data to be received */
-	uint8_t timeout       /* UART timeout */
+	void*    rx_data_ptr , /* Buffer to export data to        */
+	size_t   rx_data_size, /* Size of the data to be received */
+	uint32_t timeout       /* UART timeout */
 	);
